Change repo-branch: `main` to `master` for urls
The branch of the repo is `master` but urls are created according to default branch name `main` as mentioned <https://github.com/quarto-dev/quarto-cli/discussions/4826>. As a results, because *Edit this page** and **View source** buttons in every page contains branch names in their urls, they do not work now.

It updates `repo-branch: main` `repo-branch: master`.

Old and not working urls
-https://github.com/Lakens/statistical_inferences/edit/master/index.qmd
-https://github.com/Lakens/statistical_inferences/blob/master/index.qmd

New and working urls
-https://github.com/Lakens/statistical_inferences/edit/master/index.qmd
-https://github.com/Lakens/statistical_inferences/blob/master/index.qmd